Move component tests into expected subdirectory
This lets the linting rule find them properly.
