+++ ui/tests/integration/components/allocation-row-test.js
@@ -1,7 +1,7 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import generateResources from '../../mirage/data/generate-resources';
+import generateResources from '../../../mirage/data/generate-resources';
 import { startMirage } from 'nomad-ui/initializers/ember-cli-mirage';
 import { find, render } from '@ember/test-helpers';
 import Response from 'ember-cli-mirage/response';
