+++ ui/tests/integration/components/attributes-table-test.js
