+++ ui/tests/integration/components/gauge-chart-test.js
