+++ ui/tests/integration/components/image-file-test.js
