+++ ui/tests/integration/components/job-diff-test.js
@@ -2,7 +2,7 @@ import { findAll, find, render } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import cleanWhitespace from '../utils/clean-whitespace';
+import cleanWhitespace from '../../utils/clean-whitespace';
 import { componentA11yAudit } from 'nomad-ui/tests/helpers/a11y-audit';
 
 module('Integration | Component | job diff', function(hooks) {
