+++ ui/tests/integration/components/job-page/parts/placement-failures-test.js
