+++ ui/tests/integration/components/job-page/parts/summary-test.js
