+++ ui/tests/integration/components/plugin-allocation-row-test.js
