+++ ui/tests/integration/components/primary-metric-test.js
