+++ ui/tests/integration/components/scale-events-chart-test.js
