+++ ui/tests/integration/components/stepper-input-test.js
