+++ ui/tests/integration/components/task-log-test.js
@@ -5,7 +5,7 @@ import { find, click, render, settled } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import { componentA11yAudit } from 'nomad-ui/tests/helpers/a11y-audit';
 import Pretender from 'pretender';
-import { logEncode } from '../../mirage/data/logs';
+import { logEncode } from '../../../mirage/data/logs';
 
 const HOST = '1.1.1.1:1111';
 const allowedConnectionTime = 100;
